initial commit
Heart failure is a major concern of all time because it is a sudden event. Research took a dataset which has approximately a thousand number of samples and various attributes such as hypertension, diabetes, mental health, pacemaker, history of stroke and other various ailments. So the model tries to sort out the dataset in manner to be effective with algorithms. Numerous machine learning algorithms are experimented and out of those algorithms some algorithms such as Catboost, Adaboost, and Support vector classifier are selected which are giving high accuracy. Model came across with exploratory data analysis, data preprocessing and data visualization.